+++ src/components/BottomSheet/index.ts
@@ -1,0 +1,3 @@
+import BottomSheet from './BottomSheet';
+
+export default BottomSheet;
