+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './BottomSheet';
+export {default as BottomSheet} from './BottomSheet';
 export * from './Box';
 export * from './Button';
 export * from './CodeInput';
